feat(menu): seperate menu logic from browsing view
